Add Semantic UI's color types to the coffee constants
