--- src/pages/coffee/constants.js
@@ -1,11 +1,0 @@
-export const COFFEE_STATE_COLORS = {
-  'TOPIC_COLLECTION': 'purple',
-  'TOPIC_VOTING': 'pink',
-  'DISCUSSION': 'violet',
-}
-
-export const COFFEE_STATE_NAMES = {
-  'TOPIC_COLLECTION': 'Topic collection',
-  'TOPIC_VOTING': 'Topic voting',
-  'DISCUSSION': 'Discussion',
-}
