+++ src/pages/coffee/constants.ts
@@ -1,0 +1,17 @@
+import { SemanticCOLORS } from 'semantic-ui-react'
+
+const PURPLE: SemanticCOLORS = 'purple'
+const PINK: SemanticCOLORS = 'pink'
+const VIOLET: SemanticCOLORS = 'violet'
+
+export const COFFEE_STATE_COLORS = {
+  'TOPIC_COLLECTION': PURPLE,
+  'TOPIC_VOTING': PINK,
+  'DISCUSSION': VIOLET,
+}
+
+export const COFFEE_STATE_NAMES = {
+  'TOPIC_COLLECTION': 'Topic collection',
+  'TOPIC_VOTING': 'Topic voting',
+  'DISCUSSION': 'Discussion',
+}
